requirements.txt
streamlit==1.32.0
pandas==2.2.1
openai==1.83.0
plotly==5.19.0
python-dotenv==1.0.1
kaleido==0.2.1
openpyxl==3.1.2 